chore(clerk-js): Rename file to InvitedMembersList
